+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 #### Improvements
 * Fix #5843: Support existingJavaTypes extension in java-generator
+* Fix #5878: (java-generator) Add implements Editable for extraAnnotations
 
 ### 6.11.0 (2024-03-25)